hotfix: sql type

+++ shop-server/sql/create_tables.sql
@@ -42,8 +42,8 @@ create table shops (
     created_at date not null,
     in_vacations boolean not null,
     name varchar(255) not null,
-    latitude double not null,
-    longitude double not null,
+    latitude double precision not null,
+    longitude double precision not null,
     primary key (id)
 );
